Move path.h to namespace Common
